test: refactoring
