+++ test/src/entities.ts
@@ -6,7 +6,6 @@ dotenv.config(); // { path: path.resolve(__dirname, ".env") }
 export const adresses = {
   sepoliaTestnet: "0x52B6d10d7d865B3d4103f8809AA3521288568f46",
   zksyncMainnet: "0x52B6d10d7d865B3d4103f8809AA3521288568f46",
-  goerliTestnet: "0x52B6d10d7d865B3d4103f8809AA3521288568f46",
 };
 
 export const contracts = {
